% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameboy_plot.R
\name{gameboy_plot}
\alias{gameboy_plot}
\title{A plotting function for a custom gameboy}
\usage{
gameboy_plot(
  background = "#E0E0E0",
  case = "#293133",
  glassframe = "black",
  screen = "lightgrey",
  Abutton = "black",
  Bbutton = "black",
  start = "black",
  select = "black",
  cross = "black",
  speaker = "black",
  screentext = "Put your text here",
  labelGB = "white",
  Done_col = "white",
  R_col = "black",
  Rbackground = "white",
  ...
)
}
\arguments{
\item{background}{Color of the plot background.}

\item{case}{Color of the gameboy case.}

\item{glassframe}{Color of the glassframe.}

\item{screen}{Color of the screen.}

\item{Abutton}{Color of the A button.}

\item{Bbutton}{Color of the B button.}

\item{start}{Color of the start button.}

\item{select}{Color of the select button.}

\item{cross}{Color of the navigation cross.}

\item{speaker}{Color of the speaker points.}

\item{screentext}{Text that is shown in the screen.}

\item{labelGB}{Color of the "GAMEBOY" Label.}

\item{Done_col}{Color of the string "Done in".}

\item{R_col}{Color of the "R" character.}

\item{Rbackground}{Color of the background of the "R" character.}
}
\description{
This function allows you to custamize your own gameboy.
}
\examples{
gameboy_plot()
}
\keyword{customisation}
\keyword{gameboy}
